Move test_opt contents to test_gen
